requirements.txt
altgraph==0.17
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto==2.49.0
boto3==1.17.33
botocore==1.20.33
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==1.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
django-storages==1.11.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
fake-useragent==0.1.11
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.3
importlib-resources==5.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.11
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
parse==1.19.0
pefile==2021.5.24
Pillow==8.3.1
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pyee==8.1.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pyppeteer==0.2.5
pyquery==1.4.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.25.1
requests-html==0.10.0
s3transfer==0.3.6
selenium==3.141.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
w3lib==1.22.0
websockets==8.1
Werkzeug==2.0.1
zipp==3.4.1
